+++ packages/controller/tsconfig.json
@@ -5,6 +5,6 @@
 	],
 	"include": ["index.ts", "controller.ts", "package.json", "src/**/*.ts"],
 	"compilerOptions": {
-		"outDir": "dist",
+		"outDir": "dist/node",
 	},
 }
