+++ packages/host/tsconfig.json
@@ -5,6 +5,6 @@
 	],
 	"include": ["index.ts", "host.ts", "package.json", "src/**/*.ts", "src/**/*.js"],
 	"compilerOptions": {
-		"outDir": "dist",
+		"outDir": "dist/node",
 	},
 }
