+++ packages/lib/scripts/compile_validator.js
@@ -7,9 +7,9 @@ const standaloneCode = require("ajv/dist/standalone").default;
 /* eslint-disable node/no-sync */
 
 // Provide an empty file for message_core to load
-fs.outputFileSync(path.join(__dirname, "../dist/src/data/message_validate.js"), "");
+fs.outputFileSync(path.join(__dirname, "../dist/node/src/data/message_validate.js"), "");
 
-const { Message } = require("../dist/src/data/messages_core");
+const { Message } = require("../dist/node/src/data/messages_core");
 
 const ajv = new Ajv({
 	allowUnionTypes: true,
@@ -28,5 +28,5 @@ function write(output) {
 		`// generated by scripts/compile_validator.js\n// @ts-nocheck\n/* eslint-disable */\n${moduleCode}`
 	);
 }
-write("../dist/src/data/message_validate.js");
+write("../dist/node/src/data/message_validate.js");
 write("../src/data/message_validate.js");
