+++ packages/lib/tsconfig.json
@@ -2,7 +2,7 @@
 	"extends": "../../tsconfig.node.json",
 	"include": ["index.ts", "browser.ts", "build_mod.js", "src/**/*.ts", "src/**/*.js"],
 	"compilerOptions": {
-		"outDir": "dist",
+		"outDir": "dist/node",
 		"allowJs": true,
 	},
 }
