+++ plugins/global_chat/index.ts
@@ -6,8 +6,8 @@ export const plugin: lib.PluginDeclaration = {
 	name: "global_chat",
 	title: "Global Chat",
 	description: "Forwards chat between instances.",
-	instanceEntrypoint: "dist/plugin/instance",
-	ctlEntrypoint: "dist/plugin/control",
+	instanceEntrypoint: "dist/node/instance",
+	ctlEntrypoint: "dist/node/control",
 
 	messages: [
 		ChatEvent,
