+++ plugins/global_chat/package.json
@@ -6,7 +6,7 @@
 	"homepage": "https://github.com/clusterio/clusterio#readme",
 	"license": "MIT",
 	"repository": "clusterio/clusterio",
-	"main": "dist/plugin/index.js",
+	"main": "dist/node/index.js",
 	"scripts": {
 		"prepare": "tsc --build && webpack-cli --env production",
 		"test": "echo \"Error: run tests from root\" && exit 1"
