+++ plugins/player_auth/index.ts
@@ -15,8 +15,8 @@ export const plugin: lib.PluginDeclaration = {
 	name: "player_auth",
 	title: "Player Auth",
 	description: "Provides authentication to the cluster via logging into a Factorio server.",
-	controllerEntrypoint: "dist/plugin/controller",
-	instanceEntrypoint: "dist/plugin/instance",
+	controllerEntrypoint: "dist/node/controller",
+	instanceEntrypoint: "dist/node/instance",
 	webEntrypoint: "./web",
 	controllerConfigFields: {
 		"player_auth.code_length": {
