+++ plugins/research_sync/test/missing.js
@@ -2,5 +2,5 @@
 // required here in order to track coverage for them.
 
 "use strict";
-require("../dist/plugin/instance");
-require("../dist/plugin/controller");
+require("../dist/node/instance");
+require("../dist/node/controller");
